split out spectator to new file
