+++ common/server/game/server_abstract_participant.cpp
@@ -97,16 +97,6 @@ Server_AbstractParticipant::Server_AbstractParticipant(Server_Game *_game,
 {
 }
 
-Server_Spectator::Server_Spectator(Server_Game *_game,
-                                   int _playerId,
-                                   const ServerInfo_User &_userInfo,
-                                   bool _judge,
-                                   Server_AbstractUserInterface *_userInterface)
-    : Server_AbstractParticipant(_game, _playerId, _userInfo, _judge, _userInterface)
-{
-    spectator = true;
-}
-
 Server_AbstractParticipant::~Server_AbstractParticipant() = default;
 
 void Server_AbstractParticipant::removeFromGame()
