+++ common/server/game/server_abstract_participant.h
@@ -180,15 +180,4 @@ public:
     getInfo(ServerInfo_Player *info, Server_AbstractParticipant *recipient, bool omniscient, bool withUserInfo);
 };
 
-class Server_Spectator : public Server_AbstractParticipant
-{
-    Q_OBJECT
-public:
-    Server_Spectator(Server_Game *_game,
-                     int _playerId,
-                     const ServerInfo_User &_userInfo,
-                     bool _judge,
-                     Server_AbstractUserInterface *_handler);
-};
-
 #endif
